align metadata
